pip-requirements.txt
discord.py[voice]
gspread
imgurpython
oauth2client
pillow
pprint
pycountry
pytz
pyyaml
sqlalchemy
